Stop deleting the migrations table during testing

+++ system/Test/CIDatabaseTestCase.php
@@ -163,6 +163,8 @@ class CIDatabaseTestCase extends CIUnitTestCase
 
 				foreach ($tables as $table)
 				{
+					if ($table = $this->db->DBPrefix.'migrations') continue;
+
 					$forge->dropTable($table, true);
 				}
 			}
